Add files via upload
The software development house you work for has decided to create a prototype based on the 
game Scrabble, but with a twist, that will be used on mobile devices. They require a prototype 
from you that is designed using a Java console application. If the prototype is a success, the 
application will be developed for the mainstream mobile platforms. 
You are firstly required to prompt the user to enter in a one to start the game. If the users decide 
to play the game, prompt for the two player names.